Api key for tests (#29)
* Refactor code for setting background CSS class on body

Fixing a TODO in the code for setting the weather condition CSS class
on body.  The issue with the existing implementation is that it will
wipe out any/all existing CSS classes on body.  It wasn’t causing an
issue because no other classes were being set, but still not ideal.

* Revert "Refactor code for setting background CSS class on body"

This reverts commit 7e02a3507c530a4278a46018e8633bb9b91eafff.

* Adding support for querying OpenWeatherMap by zip code

bmorelli25/React-Weather-App#26

* Provide a friendlier name in output when searching by zip code

bmorelli25/React-Weather-App#26

* Fixing issue where tests fail because API_KEY is not defined

bmorelli25/React-Weather-App#28